+++ main.c
@@ -11,6 +11,6 @@ int 			main()
 	window = fill_wind(1920, 1080);
 	point_of_view =  fill_point(0, 0, 0);
 	view = fill_vector(0, 0, 1);
-	camera = fill_camera(point_of_view, view, (float)80, window);
+	camera = fill_camera(point_of_view, view, (double)80, window);
 return (0);
 }
